# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ushindi 2027 - Campaign Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content/assets:
    get:
      summary: List content assets
      tags: [content]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: approvalStatus, in: query, schema: { type: string } }
        - { name: countyId, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAssetListResponse"
    post:
      summary: Create content asset catalogue entry
      tags: [content]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentAssetInput"
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAsset2"
  /content/assets/{id}:
    get:
      summary: Get content asset
      tags: [content]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Asset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAssetDetail"
    patch:
      summary: Update content asset metadata
      tags: [content]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentAssetInput"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAsset2"
  /content/assets/{id}/approve:
    post:
      summary: Approve content asset
      tags: [content]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAsset2"
  /content/assets/{id}/reject:
    post:
      summary: Reject content asset
      tags: [content]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAsset2"
  /content/assets/{id}/download:
    post:
      summary: Record download and get object path
      tags: [content]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                purpose: { type: string }
      responses:
        "200":
          description: Download info
          content:
            application/json:
              schema:
                type: object
                properties:
                  objectPath: { type: string }
                  title: { type: string }
  /content/assets/{id}/history:
    get:
      summary: Asset download history
      tags: [content]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Download records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DownloadRecord"

  # ── EVENTS MANAGEMENT ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ContentAsset2:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        objectPath: { type: string }
        mimeType: { type: string }
        publishingRights: { type: string }
        approvalStatus: { type: string }
        currentVersion: { type: integer }
        downloadCount: { type: integer }
        countyId: { type: string }
        language: { type: string }
        expiresAt: { type: string }
        createdAt: { type: string }
    ContentAssetDetail:
      allOf:
        - $ref: "#/components/schemas/ContentAsset2"
        - type: object
          properties:
            versions:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  version: { type: integer }
                  objectPath: { type: string }
                  changeNote: { type: string }
                  createdAt: { type: string }
            downloadCount: { type: integer }
    ContentAssetListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ContentAsset2"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    ContentAssetInput:
      type: object
      required: [title, category, objectPath]
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        objectPath: { type: string }
        mimeType: { type: string }
        publishingRights: { type: string }
        countyId: { type: string }
        language: { type: string }
        expiresAt: { type: string }
    DownloadRecord:
      type: object
      properties:
        id: { type: string }
        assetId: { type: string }
        downloadedByEmail: { type: string }
        purpose: { type: string }
        createdAt: { type: string }

    # ── EVENTS SCHEMAS ───────────────────────────────────────────────────────
...