# lib/api-zod/src/generated/types/contentAssetInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Ushindi 2027 - Campaign Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ContentAssetInput {
  title: string;
  description?: string;
  category: string;
  objectPath: string;
  mimeType?: string;
  publishingRights?: string;
  countyId?: string;
  language?: string;
  expiresAt?: string;
}