# lib/api-zod/src/generated/types/downloadRecord.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Ushindi 2027 - Campaign Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadRecord {
  id?: string;
  assetId?: string;
  downloadedByEmail?: string;
  purpose?: string;
  createdAt?: string;
}